Move security manager override to its own module
