+++ tests/auth/managers/fab/test_fab_auth_manager.py
@@ -24,7 +24,7 @@ import pytest
 from airflow import AirflowException
 from airflow.auth.managers.fab.fab_auth_manager import FabAuthManager
 from airflow.auth.managers.fab.models import User
-from airflow.auth.managers.fab.security_manager_override import FabAirflowSecurityManagerOverride
+from airflow.auth.managers.fab.security_manager.override import FabAirflowSecurityManagerOverride
 from airflow.www.security import ApplessAirflowSecurityManager
